+++ frontend/src/layouts/pages/mypages-sections/statistics/export-statistics-button/export-statistics-modal.component.tsx
@@ -1,27 +1,22 @@
 'use client';
 
-import { ChevronDown, ChevronUp } from 'lucide-react';
 import {
   Button,
   Checkbox,
-  Divider,
   FormErrorMessage,
   FormLabel,
   Modal,
   RadioButton,
   Select,
-  Accordion,
   FormControl,
-  SearchField,
-  NavigationBar,
 } from '@sk-web-gui/react';
 import { useTranslation } from 'react-i18next';
-import { ChangeEvent, useEffect, useMemo, useState } from 'react';
+import { ChangeEvent, useCallback, useEffect, useState } from 'react';
 import { CustomDatePicker } from './custom-date-picker.component';
-import { useCheckboxTree } from './use-checkbox-tree';
+import { DatePeriodToggle } from './date-period-toggle.component';
+import { FacilitiesAccordion } from './facilities-accordion.component';
 import { useApi } from '@services/api-service';
 import { User } from '@interfaces/user';
-import { getCategoryFromInstalledBaseType, getEventCategory } from '@utils/facility';
 import { Category } from '@interfaces/measurement-data';
 import {
   DatePeriod,
@@ -34,7 +29,6 @@ import {
 } from './date-picker.util';
 
 const categories = [Category.DISTRICT_HEATING, Category.ELECTRICITY];
-const datePeriods = ['year', 'month', 'day'];
 const timeIntervals = ['hour', 'quarter'];
 
 export interface ExportModalData {
@@ -67,9 +61,10 @@ export const ExportStatisticsModal = ({
   initialTimeResolution,
   initialFacilityId,
 }: ExportStatisticsModalProps) => {
-  const [searchValue, setSearchValue] = useState('');
-  const [isSearchDirty, setIsSearchDirty] = useState(false);
   const [temperatureIncluded, setTemperatureIncluded] = useState(false);
+  const [checkedFacilities, setCheckedFacilities] = useState<Set<string>>(new Set());
+  const [checkedFacilitiesCount, setCheckedFacilitiesCount] = useState(0);
+  const [totalFacilitiesCount, setTotalFacilitiesCount] = useState(0);
   const [category, setCategory] = useState<Category>(initialCategory ?? categories[0]);
   const [fromDate, setFromDate] = useState(initialFromDate ?? '');
   const [toDate, setToDate] = useState(initialToDate ?? '');
@@ -79,60 +74,17 @@ export const ExportStatisticsModal = ({
   );
   const { t } = useTranslation(['statistics']);
 
+  const handleFacilitySelectionChange = useCallback((items: Set<string>, count: number, total: number) => {
+    setCheckedFacilities(items);
+    setCheckedFacilitiesCount(count);
+    setTotalFacilitiesCount(total);
+  }, []);
+
   const { data: user } = useApi<User>({
     method: 'get',
     url: '/me',
     queryKey: ['user'],
   });
-
-  const facilitiesGroupedByAddresses = useMemo(() => {
-    const filtered = user?.facilities?.filter((f) => getCategoryFromInstalledBaseType(f.type) === category) ?? [];
-    const groupMap = new Map<string, string[]>();
-    for (const facility of filtered) {
-      const address = facility.address?.street ?? '';
-      if (!groupMap.has(address)) groupMap.set(address, []);
-      groupMap.get(address)!.push(facility.facilityId ?? '');
-    }
-    return Array.from(groupMap.entries()).map(([address, facilities]) => ({ address, facilities }));
-  }, [user, category]);
-
-  const filteredGroups = useMemo(() => {
-    const query = searchValue.toLowerCase().trim();
-    if (!query) return facilitiesGroupedByAddresses;
-
-    const matchesFacilityQuery = (facilityId: string): boolean => {
-      if (facilityId.toLowerCase().includes(query)) return true;
-      const facility = user?.facilities?.find((f) => f.facilityId === facilityId);
-      const categoryKey = getCategoryFromInstalledBaseType(facility?.type);
-      const typeLabel = categoryKey ? t(`statistics:exportModal.category.${categoryKey}`) : (facility?.type ?? '');
-      return typeLabel.toLowerCase().includes(query);
-    };
-
-    return facilitiesGroupedByAddresses.flatMap((group) => {
-      if (group.address.toLowerCase().includes(query)) return [group];
-      const matchingFacilities = group.facilities.filter(matchesFacilityQuery);
-      return matchingFacilities.length > 0 ? [{ ...group, facilities: matchingFacilities }] : [];
-    });
-  }, [searchValue, facilitiesGroupedByAddresses, user, t]);
-
-  const checkboxGroups = useMemo(
-    () => facilitiesGroupedByAddresses.map((group) => ({ key: group.address, items: group.facilities })),
-    [facilitiesGroupedByAddresses]
-  );
-
-  const {
-    checkedItems,
-    checkedCount: checkedFacilitiesCount,
-    totalCount: totalFacilitiesCount,
-    allChecked: allFacilitiesChecked,
-    noneChecked: allFacilitiesUnchecked,
-    toggleItem: toggleFacility,
-    toggleGroup: toggleAddress,
-    toggleAll: toggleAllFacilities,
-    groupStates: addressStates,
-    isItemChecked: isFacilityChecked,
-    reset: resetFacilities,
-  } = useCheckboxTree(checkboxGroups);
 
   useEffect(() => {
     if (!show) return;
@@ -144,24 +96,8 @@ export const ExportStatisticsModal = ({
     setFromDate(formatDateForInputType(initialFromDate ?? '', inputType));
     setToDate(formatDateForInputType(initialToDate ?? '', inputType));
     setTemperatureIncluded(false);
-    setSearchValue('');
-    setIsSearchDirty(false);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [show]);
-
-  useEffect(() => {
-    if (!show || !user || !initialFacilityId) return;
-    const facility = user.facilities?.find((f) => f.facilityId === initialFacilityId);
-    const address = facility?.address?.street;
-    resetFacilities(address ? [`${address}::${initialFacilityId}`] : []);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [show, user, initialFacilityId]);
-
-  useEffect(() => {
-    resetFacilities([]);
-    setTimeInterval('hour');
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [category]);
 
   const datePickerType = getDatePickerType(datePeriod);
 
@@ -172,7 +108,7 @@ export const ExportStatisticsModal = ({
   const aggregation = aggregationByPeriod[datePeriod] ?? timeInterval;
 
   const handleExport = () => {
-    const selectedFacilities = Array.from(checkedItems).map((key) => {
+    const selectedFacilities = Array.from(checkedFacilities).map((key) => {
       const [address, facilityId] = key.split('::');
       return { facilityId, address };
     });
@@ -228,25 +164,7 @@ export const ExportStatisticsModal = ({
           <div className="lg:pt-0 pt-16 lg:justify-end justify-center lg:flex-[0_0_auto] w-full lg:w-auto">
             <div className="flex flex-col gap-8 w-full lg:pt-0 pt-16">
               <FormLabel>{t('statistics:exportBy')}</FormLabel>
-              <NavigationBar
-                className="!py-6 bg-tertiary-surface flex justify-around"
-                size="md"
-                data-cy="export-modal-date-toggle"
-              >
-                {datePeriods.map((interval) => (
-                  <NavigationBar.Item key={`time-interval-${interval}`} className="lg:w-auto w-full !p-0 !m-0">
-                    <Button
-                      className="lg:w-auto w-full !h-[12px] !py-0"
-                      size="sm"
-                      inverted={datePeriod === interval}
-                      onClick={() => handleDatePeriodChange(interval as DatePeriod)}
-                      data-cy={`export-modal-date-toggle-${interval}-button`}
-                    >
-                      {t(`statistics:${interval}`)}
-                    </Button>
-                  </NavigationBar.Item>
-                ))}
-              </NavigationBar>
+              <DatePeriodToggle value={datePeriod} onChange={handleDatePeriodChange} />
             </div>
           </div>
 
@@ -289,98 +207,13 @@ export const ExportStatisticsModal = ({
             </FormControl>
           )}
           <div className="w-full">
-            <Accordion className="mr-0">
-              <Accordion.Item className="mr-0">
-                <Accordion.Item.Header
-                  className="flex flex-row items-start"
-                  data-cy="export-facilities-accordion-header"
-                >
-                  <Accordion.Item.Title className="flex flex-col items-start gap-6 flex-1">
-                    <div className="text-content text-label-medium">{t('statistics:exportModal.facilities')}</div>
-                    <div className="text-content text-small font-normal">
-                      {t('statistics:exportModal.selection', {
-                        count: checkedFacilitiesCount,
-                        total: totalFacilitiesCount,
-                      })}
-                    </div>
-                  </Accordion.Item.Title>
-                  <Accordion.Item.Button>
-                    {(open: boolean) => (open ? <ChevronUp /> : <ChevronDown />)}
-                  </Accordion.Item.Button>
-                </Accordion.Item.Header>
-                <Accordion.Item.Content className="mb-8 mr-0 !pr-0 ">
-                  <div className="mr-0 pr-0 flex flex-col items-start self-stretch gap-16">
-                    <div className="pb-8 flex flex-col items-start self-stretch gap-8">
-                      <SearchField
-                        placeholder={t('statistics:exportModal.searchFacilities')}
-                        className="py-8 pr-4 pl-4 flex-1 flex items-center self-stretch gap-8"
-                        value={searchValue}
-                        showSearchButton={isSearchDirty}
-                        onChange={(event: ChangeEvent<HTMLInputElement>) => {
-                          setSearchValue(event.target.value);
-                          setIsSearchDirty(true);
-                        }}
-                        onSearch={() => {
-                          setIsSearchDirty(false);
-                        }}
-                        onReset={() => {
-                          setSearchValue('');
-                          setIsSearchDirty(false);
-                        }}
-                      />
-                    </div>
-                    <div className="pl-8 flex flex-col gap-12 items-start self-stretch">
-                      <Checkbox
-                        checked={!allFacilitiesUnchecked}
-                        indeterminate={!allFacilitiesChecked && !allFacilitiesUnchecked}
-                        onChange={toggleAllFacilities}
-                        className="py-8 pr-2 flex items-center self-stretch"
-                        data-cy="export-facilities-select-all"
-                      >
-                        {t('statistics:exportModal.selectAll')}
-                      </Checkbox>
-                      <div className="pl-16 flex flex-col gap-12 items-start self-stretch">
-                        {filteredGroups.map((group) => (
-                          <div key={group.address} className="flex flex-col gap-4 items-start self-stretch">
-                            <Checkbox
-                              checked={addressStates.get(group.address)?.checked ?? false}
-                              indeterminate={addressStates.get(group.address)?.indeterminate ?? false}
-                              onChange={() => toggleAddress(group.address, group.facilities)}
-                              className="flex self-stretch py-8 pr-2 text-small font-normal"
-                            >
-                              {group.address}
-                            </Checkbox>
-                            <div className="pl-16 flex flex-col gap-4 items-start self-stretch">
-                              {group.facilities.map((facilityId) => {
-                                const facility = user?.facilities?.find(
-                                  (f) => f.facilityId === facilityId && f.type !== 'Elhandel'
-                                );
-                                const categoryKey = getEventCategory(facility?.type);
-                                const typeLabel = categoryKey
-                                  ? t(`statistics:exportModal.category.${categoryKey}`)
-                                  : (facility?.type ?? facilityId);
-                                const displayName = `${typeLabel} (${facilityId})`;
-                                return (
-                                  <Checkbox
-                                    key={facilityId}
-                                    checked={isFacilityChecked(group.address, facilityId)}
-                                    onChange={() => toggleFacility(group.address, facilityId)}
-                                    className="py-8 pr-2 self-stretch"
-                                  >
-                                    {displayName}
-                                  </Checkbox>
-                                );
-                              })}
-                            </div>
-                          </div>
-                        ))}
-                      </div>
-                    </div>
-                  </div>
-                </Accordion.Item.Content>
-                <Divider />
-              </Accordion.Item>
-            </Accordion>
+            <FacilitiesAccordion
+              show={show}
+              user={user}
+              category={category}
+              initialFacilityId={initialFacilityId}
+              onSelectionChange={handleFacilitySelectionChange}
+            />
           </div>
           <div className="w-full">
             <Checkbox checked={temperatureIncluded} onChange={() => setTemperatureIncluded((prev) => !prev)}>
